Migrate to Swift 6.2
This "migration" sets the default actor isolation to the MainActor.

I believe this is a valid setting in this context. Stinsen is a UI library only,
and running it off the MainActor will introduce (subtle) bugs, if changes are
published and help off the main thread.

In addition, there was a state update within a state update, so I deferred assigning
"activeTab" in TabChild into a new task (i.e. to another next run loop iteration).

These changes are not meant to upgrade the Stinsen implementation as a whole — rather to bring
Stinsen ready-to-use for Swift 6.2.